size is almost correct

+++ proto/encode_gogo.go
@@ -53,7 +53,7 @@ func (o *Buffer) enc_ref_int32(p *Properties, base structPointer) error {
 	}
 	x := refWord32_Get(v)
 	o.buf = append(o.buf, p.tagcode...)
-	p.valEnc(o, uint64(int32(x)))
+	p.valEnc(o, uint64(x))
 	return nil
 }
